Use real physical control net device in /var/emulab/boot/controlif .
Somehow we had decided to to put br-ex into this file.  That is
technically correct, but not from slothd's perspective, since virtual
interfaces can plug into br-ex but not inject traffic from the host
br-ex interface... so we miss that traffic.  And we can't miss it.

(We still stash br-ex into /var/run/cnet... creating this mixed message
is not ideal, but we have to fix the slothd control net data collection
issue, which means lying to /var/emulab/boot/controlif.)